Refactor: main에서 기존 ray.remote process 추출 및 레거시 명명 #24
